changeOfCombat to float

+++ src/main/java/Enum/TerrainFeatures.java
@@ -1,6 +1,7 @@
 package Enum;
 
 public enum TerrainFeatures {
+    // TODO type of changeOfCombat
     FLOODPLAINS("floodplains", 2, 0, 0, -33/100, 1, new String[]{"wheat", "sugar"}),
     FOREST("forest", 1, 1, 0, 25/100, 2, new String[]{"uranium", "gazelle", "wool", "color", "silk"}),
     ICE("ice", 0,0,0, 0, 10000, null),
@@ -71,7 +72,7 @@ public enum TerrainFeatures {
         return changeOfCombat;
     }
 
-    public void setChangeOfCombat(int changeOfCombat) {
+    public void setChangeOfCombat(float changeOfCombat) {
         this.changeOfCombat = changeOfCombat;
     }
